Fix prefabs READ DESCRIPTION! :arrow_heading_down:
If you want an item to be flammable, give it a flammable script on the
object that has a collider. The collider is needed to scale the fire.
For all existing scenery prefabs, the flammable script belongs on the
child object. (Ex: add the script to pcylinder5 instead of barrel)